Bump github.com/opencontainers/image-spec to v1.0.2
Fix for GHSA-77vh-xpmg-72qh

Vulnerable versions: < 1.0.2
Patched version: 1.0.2

`Impact`
In the OCI Image Specification version 1.0.1 and prior, manifest
and index documents are not self-describing and documents with
a single digest could be interpreted as either a manifest or an
index.

`Patches`
The Image Specification will be updated to recommend that both
manifest and index documents contain a mediaType field to
identify the type of document.

`Workarounds`
Software attempting to deserialize an ambiguous document may
reject the document if it contains both “manifests” and “layers”
fields or “manifests” and “config” fields.

`References`
GHSA-mc8v-mgrf-8f4m

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>