proxy: call maybeReexec() up front
Commands like `skopeo copy` do this first depending on the image names
involved in the operation, but we don't have the luxury of having access
to that information up-front.  Always call maybeReexec(), just in case
it's required later on.

Add a stub for maybeReexec() to cmd/skopeo/unshare.go for non-Linux, as
is already done for reexecIfNecessaryForImages().

Fixes #2563

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>