integration: default to "registry" binary
The "registry-v2" binary currently being used is present in the
container image but it can't be found on local environments.

The docker-distribution provides the `registry` binary which does work
for the integration tests.

This commit switches to preferring the `registry-v2` binary and defaults
to `registry`. `make test-integration-local` passes with this change
and docker-distribution installed.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>