skopeo: add `--reject-insecure-policy`
In bootc, we want the ability to assert that signature verification is
enforced.

Add a new top-level `--reject-insecure-policy` switch (at the same level
as the existing `--insecure-policy` switch). When passed, we use the
new `SetRejectPolicy()` method to ensure that signature verification
is enforced.

Part of https://github.com/containers/skopeo/issues/1829.

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>